chore: add error check to KillNTSC (#8441)

+++ master/internal/command/command_service.go
@@ -3,6 +3,7 @@ package command
 import (
 	"context"
 	"fmt"
+	"strings"
 	"sync"
 
 	"github.com/sirupsen/logrus"
@@ -231,7 +232,7 @@ func (cs *CommandService) KillNTSC(id string, taskType model.TaskType) (*Command
 
 	if !completed {
 		err = task.DefaultService.Signal(c.allocationID, task.KillAllocation, "user requested kill")
-		if err != nil {
+		if err != nil && !strings.Contains(err.Error(), "not found") {
 			return nil, fmt.Errorf("failed to kill allocation: %w", err)
 		}
 	}
